move /api/duels functionality into /api/battles
