+++ server/battles/battleRoutes.js
@@ -2,4 +2,6 @@ var battleController = require('./battleController.js');
 
 module.exports = function (app) {
   app.post('/checkvalidbattleroom', battleController.checkvalidbattleroom);
+  app.post('/getbattle', battleController.getBattle);
+  app.post('/attemptbattle', battleController.attemptBattle);
 };
